导入

+++ src/main/java/com/xingrongjinfu/commodity/register/model/Register.java
@@ -1,6 +1,5 @@
 package com.xingrongjinfu.commodity.register.model;
 
-import java.math.BigDecimal;
 import java.util.Date;
 
 public class Register {
@@ -13,9 +12,9 @@ public class Register {
     public Double inPrice;//进价（单位为分）
     public Double vipPrice;//会员价
     public Integer discount;//是否折扣，1是，-1否
-    public BigDecimal stockNum;//库存
-    public Integer upperLimit;//库存上限
-    public Integer lowerLimit;//库存下限
+    public Double stockNum;//库存
+    public Double upperLimit;//库存上限
+    public Double lowerLimit;//库存下限
     
     public String commodityCode;//商品编码
     public String supplierName;//供应商名
@@ -185,22 +184,22 @@ public class Register {
     public void setCommodityStatus(String commodityStatus) {
         this.commodityStatus = commodityStatus;
     }
-    public BigDecimal getStockNum() {
+    public Double getStockNum() {
         return stockNum;
     }
-    public void setStockNum(BigDecimal stockNum) {
+    public void setStockNum(Double stockNum) {
         this.stockNum = stockNum;
     }
-    public Integer getUpperLimit() {
+    public Double getUpperLimit() {
         return upperLimit;
     }
-    public void setUpperLimit(Integer upperLimit) {
+    public void setUpperLimit(Double upperLimit) {
         this.upperLimit = upperLimit;
     }
-    public Integer getLowerLimit() {
+    public Double getLowerLimit() {
         return lowerLimit;
     }
-    public void setLowerLimit(Integer lowerLimit) {
+    public void setLowerLimit(Double lowerLimit) {
         this.lowerLimit = lowerLimit;
     }
     public Integer getDiscount() {
